+++ locale/ti/index.js
@@ -4,5 +4,5 @@ module.exports =
 	long: require('./long.json'),
 	short: require('./short.json'),
 	narrow: require('./narrow.json'),
-	plural: require('./plural')
+	quantify: require('./quantify')
 }